Refactor backend for RadPair Beta: Updated Dockerfile to use new app module and removed deprecated server_radpair_v2.py. Enhanced server_radpair.py with accumulative transcript functionality and integrated new German punctuation commands. The polish step has been removed for faster transcription. Updated WebSocket handling for improved performance and user experience.
